\name{exampleData}
\docType{data}
\alias{exampleData}
\title{A small example qPCR data set}
\description{Amplification measurements from the OpenArray qPCR platform
  for the let-7 family of miRNAs across three samples.}
\usage{
data(exampleData)
}
\format{A data.frame containing the raw amplification data.}

\keyword{datasets}
\examples{
 data(exampleData)
 show(exampleData)
}
